Разбиение написанного кода на модули
