+++ js/data.js
@@ -1,0 +1,14 @@
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',];
+
+const NAMES = ['Игорь', 'Степан', 'Иосиф', 'Петрушка322', 'Опасный228', 'Леха', 'Светлана Викторовна', 'Неопознанный Енот', 'Стиральная машинка BOSH',];
+
+const DESCRIPTIONS = ['Отдыхаю', 'Загораю', 'Программирую', 'Занят', 'Крутое фото', 'Завидуйте', 'Я на море', 'Было весело'];
+
+
+
+export {MESSAGES, NAMES, DESCRIPTIONS};
