+++ js/task 2/task2.js
@@ -1,0 +1,20 @@
+let random = function (a, b) {
+    if (a < 0) {
+      a = 0;
+    }
+  
+    return Math.random() * (b - a) + a;
+  
+  };
+  
+  let commentLength = function (string, maxLength) {
+  
+    if (string.length > maxLength) {
+      string = '';
+    } else {
+    }
+  
+  }
+  
+  random(0, 15);
+  commentLength('Второе домашнее задание успешно выполнено', 140);
